Move internal items to .Model.
Start abstracting the print to console logic.
