+++ ConsoleHero/GlobalUsings.cs
@@ -1,2 +1,3 @@
 ﻿global using static System.Console;
 global using System.Drawing;
+global using ConsoleHero.Model;
