+++ ConsoleHero/Model/ILineComponent.cs
@@ -1,0 +1,6 @@
+﻿namespace ConsoleHero.Model;
+
+internal interface ILineComponent
+{
+    internal Color Color { get; }
+}
