+++ ConsoleHero/Model/InputModifier.cs
@@ -1,14 +1,4 @@
-﻿namespace ConsoleHero;
-
-internal interface ILineComponent
-{
-    internal Color Color { get; }
-}
-
-internal readonly struct InputPlaceholder(Color color) : ILineComponent
-{
-    Color ILineComponent.Color { get; } = color;
-}
+﻿namespace ConsoleHero.Model;
 
 internal readonly struct InputModifier(Func<string, string> modifier, Color? color = null) : ILineComponent
 {
