+++ ConsoleHero/Model/InputPlaceholder.cs
@@ -1,0 +1,6 @@
+﻿namespace ConsoleHero.Model;
+
+internal readonly struct InputPlaceholder(Color color) : ILineComponent
+{
+    Color ILineComponent.Color { get; } = color;
+}
