+++ ConsoleHero/Model/ParagraphLine.cs
@@ -1,0 +1,6 @@
+﻿namespace ConsoleHero.Model;
+
+internal readonly struct ParagraphLine()
+{
+    internal List<ILineComponent> Components { get; } = [];
+}
