--- ConsoleHero/ParagraphLine.cs
@@ -1,6 +1,0 @@
-﻿namespace ConsoleHero;
-
-internal class ParagraphLine()
-{
-    internal List<ILineComponent> Components { get; } = [];
-}
